+++ saveup/server/src/main/java/com/junlongk/server/services/AccountService.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -17,7 +18,7 @@ public class AccountService {
     @Autowired
     private AccountRepository accountRepo;
 
-    public String addAccount(String accountName, float balance, String userId) {
+    public String addAccount(String accountName, BigDecimal balance, String userId) {
         String accountId = UUID.randomUUID().toString();
 
         Account account = new Account(accountId, accountName,
@@ -32,12 +33,12 @@ public class AccountService {
         return accountRepo.getAccountsByUserId(userId);
     }
 
-    public boolean modifyBalance(float amount, String accountId) {
+    public boolean modifyBalance(BigDecimal amount, String accountId) {
         return accountRepo.modifyBalance(amount, accountId);
     }
 
     @Transactional(rollbackFor = TransferException.class)
-    public void updateBalanceByTransfer(float amount,
+    public void updateBalanceByTransfer(BigDecimal amount,
                                  String fromAccountId,
                                  String toAccountId) throws TransferException {
         boolean statusOfFromAcct = accountRepo.updateFromBalance(amount, fromAccountId);
